+++ designsystem/src/main/java/com/everli/designsystem/helper/DpExtensions.kt
@@ -1,6 +1,0 @@
-package com.everli.designsystem.helper
-
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.ui.unit.Dp
-
-fun Dp.toRoundedShape() = RoundedCornerShape(this)