+++ designsystem/src/test/java/com/everli/designsystem/helper/EverliColorExtensionsTest.kt
@@ -39,6 +39,7 @@ internal class EverliColorExtensionsTest : FunSpec(
       EverliColor.GOOGLE to EverliColors.Google,
       EverliColor.APPLE to EverliColors.Apple,
       EverliColor.BLIK to EverliColors.Blik,
+      null to EverliColors.Transparent,
     ).forEach { (everliColor, color) ->
       test("$everliColor should be converted to $color") {
         everliColor.toColor() shouldBe color
@@ -74,11 +75,12 @@ internal class EverliColorExtensionsTest : FunSpec(
       EverliColor.GOOGLE to R.color.google,
       EverliColor.APPLE to R.color.apple,
       EverliColor.BLIK to R.color.blik,
+      null to R.id.none,
     ).forEach { (everliColor, resourceId) ->
       test("$everliColor should be converted to $resourceId") {
         everliColor.toResourceId() shouldBe resourceId
       }
     }
 
-  }
+  },
 )